Backend/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
werkzeug
Flask-Cors
requests
Flask-JWT-Extended
pydantic
pytest
pytest-flask
flask-limiter